Fail the build if image dependencies don't unmarshal
